Switch to 32-bit symbol ID, for 24-byte Monomial object.

+++ cpp/lib_moment/integer_types.h
@@ -40,7 +40,7 @@ namespace Moment {
     /**
      * For enumerating symbolic algebra elements
      */
-    using symbol_name_t = int64_t;
+    using symbol_name_t = int32_t;
 
     /**
      * The maximum length of operator sequence before heap allocations are required
